pyproject.toml
[tool.pyright]
reportPrivateImportUsage = false
reportCallIssue = false
reportMissingImports = false
reportArgumentType = false
reportUnknownParameterType = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownArgumentType = false

[tool.pylint]
disable = [
    "import-error",
    "no-member"
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-support-agent"
version = "1.0.0"
description = "AI support agent with Redmine integration and RAG knowledge base"
authors = [{name = "Support Team"}]
dependencies = [
    "streamlit>=1.28.0",
    "langgraph>=0.0.40",
    "langchain>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "langchain-community>=0.0.20",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "plotly>=5.15.0",
    "arize-phoenix>=3.0.0",
    "openinference-instrumentation-langchain>=0.1.43",
    "google-generativeai>=0.3.0",
    "tiktoken>=0.5.0",
    "python-dotenv>=1.0.0",
    "faiss-cpu>=1.7.4",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.2",
    "python-docx>=1.1.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "requests>=2.31.0"
] 